This is second eda projects of market sales
This first section is typically one paragraph designed to introduce the general ideas that are to be explored in greater detail in the rest of the analysis. The introduction may also include some general background on the topic. This provides the reader with context for the analysis and any important information needed to aid their understanding.